change ck-editor file place, put config file outside
