+++ src/components/controls/preview.vue
@@ -534,8 +534,7 @@
         { name: 'Phoenix', language: 'Elixir' }
     ];
 
-    // for km-ck-editor
-    const ckText= ref("");
+
    
 </script>
